Refactor for filenames and HTTP Verbs.
