+++ deployments/deployment-routes.js
@@ -4,11 +4,11 @@ const getDeployments = require('./controllers/deployments');
 /**
  * Retrieve the list of deployment areas of interest.
  */
-router.route('/').get(getDeployments.find);
+router.route('/').get(getDeployments.getAll);
 
 /**
  * Retrieve meta-data for a given deployment.
  */
-router.route('/:deployment').get(getDeployments.findOne);
+router.route('/:deployment').get(getDeployments.get);
 
 module.exports = router;
