+++ package.json
@@ -4,7 +4,6 @@
   "description": "OpenMapKit Server is the lightweight server component of OpenMapKit that handles the collection and aggregation of OpenStreetMap and OpenDataKit data.",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js",
     "test": "tape test/**/**/*.test.js | faucet"
   },
   "repository": {
